# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for Kuro API",
        "title": "Shop E-commerce ERP",
        "contact": {
            "name": "API Support",
            "email": "hoaiphong01012002@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/accounts/update/adjustment": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Create the inventory's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Create Inventory Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Inventory ID",
                        "name": "_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Inventory adjustment",
                        "name": "adjustment",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/budgets/create": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Create the budget's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Create Budget Information",
                "parameters": [
                    {
                        "description": "Budget data",
                        "name": "Budget",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/budgets_domain.Input"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/budgets/delete": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Delete the budget's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Delete Budget Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Budget ID",
                        "name": "_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/contracts/create": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Create the contract's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "Create Contract Information",
                "parameters": [
                    {
                        "description": "Contract data",
                        "name": "Contract",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contracts_domain.Input"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/contracts/get/_id": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Deletes the contract's information by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "Get Contract Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/departments/delete": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Deletes the department's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Delete Department Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department ID",
                        "name": "_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/departments/get/_id": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Retrieves the department's information name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Get Department Information By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/employees/create": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Create the employee's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Create Employee Information",
                "parameters": [
                    {
                        "description": "Employee data",
                        "name": "attendance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/employees_domain.Input"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/leave_requests/create": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Create the leave request's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leave Request"
                ],
                "summary": "Create Leave Request Information",
                "parameters": [
                    {
                        "description": "Leave Request data",
                        "name": "LeaveRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/leave_request_domain.Input"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/performance_reviews/update": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Update the performance review's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Performance Review"
                ],
                "summary": "Update Performance Review Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Performance Review  ID",
                        "name": "_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Performance Review data",
                        "name": "PerformanceReview",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/performance_review_domain.Input1"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/salaries/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create new salary",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Salary"
                ],
                "summary": "Create salary",
                "parameters": [
                    {
                        "description": "Salary data",
                        "name": "Salary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/salary_domain.Salary"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v2/performance_reviews/update": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Update the performance review's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Performance Review"
                ],
                "summary": "Update Performance Review Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Performance Review  ID",
                        "name": "_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Performance Review data",
                        "name": "PerformanceReview",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/performance_review_domain.Input1"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "budgets_domain.Input": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "budget_name": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "product_category": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "contracts_domain.Input": {
            "type": "object",
            "properties": {
                "contract_type": {
                    "description": "Example: \"Full-Time\", \"Part-Time\", \"Temporary\"",
                    "type": "string"
                },
                "employee": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "description": "Example: \"Active\", \"Expired\"",
                    "type": "string"
                }
            }
        },
...
        "employees_domain.Input": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Bình Thuận"
                },
                "avatarURL": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string",
                    "example": "01/01/2002"
                },
                "dayOfWork": {
                    "type": "string",
                    "example": "30/8/2024"
                },
                "department": {
                    "type": "string",
                    "example": "marketing"
                },
                "email": {
                    "type": "string",
                    "example": "hoaiphong01012002@gmail.com"
                },
                "firstName": {
                    "type": "string",
                    "example": "Ngô"
                },
                "gender": {
                    "type": "string",
                    "example": "Nam"
                },
                "lastName": {
                    "type": "string",
                    "example": "Hoài Phong"
                },
                "phone": {
                    "type": "string",
                    "example": "0329245971"
                },
                "role": {
                    "type": "string",
                    "example": "developer"
                }
            }
        },
...
        "leave_request_domain.Input": {
            "type": "object",
            "properties": {
                "employee": {
                    "type": "string",
                    "example": "admin@admin.com"
                },
                "end_date": {
                    "type": "string",
                    "example": "20/07/2024"
                },
                "leave_type": {
                    "description": "Example: \"Sick Leave\", \"Annual Leave\", \"Unpaid Leave\"",
                    "type": "string",
                    "example": "Sick Leave"
                },
                "start_date": {
                    "type": "string",
                    "example": "20/07/2024"
                },
                "status": {
                    "description": "Example: \"Approved\", \"Pending\", \"Rejected\"",
                    "type": "string",
                    "example": "Approved"
                }
            }
        },
        "performance_review_domain.Input1": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "string"
                },
                "employee": {
                    "type": "string"
                },
                "performance_score": {
                    "type": "integer"
                },
                "review_date": {
                    "type": "string"
                },
                "reviewer": {
                    "type": "string"
                }
            }
        },
...
        "salary_domain.Salary": {
            "type": "object",
            "properties": {
                "base_salary": {
                    "type": "number"
                },
                "bonus": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "deductions": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "net_salary": {
                    "type": "number"
                },
                "role_id": {
                    "type": "string"
                },
                "unit_currency": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
...
}